For clubb:ticket:788:
I made sure `wp3_pr3_rhs` was initialized when `l_use_wp3_pr3 = F`, so that zero is output to stats, rather than an uninitialized variable. This should resolve the failures of the Nightly Uninitialized Variables Test (build:15959) and the Nightly OpenMP Test (build:15953) last night.

This commit is BIT_CHANGING because a variable which was output as uninitialized (usually NaN) before is now output as zero. However, answers remain unchanged as long as `l_use_wp3_pr3 = F`.

It is worth noting that clubb:r8198 also caused the Nightly Bin Diff Regression Test (build:15948) to fail because it added a stats output and was not marked with a BIT_CHANGING tag. Including the tag in this commit will resolve that failure, and we should see a lovely shade of spring green tomorrow morning.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8199 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 88ee4f1f9682fbaf959972a78008162165f1b04f