I have updated subroutine advance_clubb_core so that pdf_params_zm is now passed into and out of it as an intent(inout) variables.  This enables it to be preserved from timestep to timestep.  This also removes the flag l_call_pdf_closure_twice from being blacklisted when ipdf_post_advance_fields is used as the ipdf_call_placement option.
The results in GrADS output files from this revision are an exact match
to the results in GrADS output files from the previous revision of this
branch for all cases (with the default configuration of
ipdf_call_placement = ipdf_pre_advance_fields).

Larson Group Internal Trac System Reference: cam:ticket:87.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/branches/reorder_advance_clubb_core@8314 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 7326a7a09d64f47ad1738b2f42b3752f9ccc1bcf