I renamed the subroutine pack_pdf_params found in setup_clubb_pdf_params.F90 as subroutine pack_hydromet_pdf_params.
I took out the ifdef CLUBB_CAM preprocessor flags around some declarations
and subroutines in pdf_parameter_module.F90 because they are also going to
be called by WRF-CLUBB (and possibly SAM-CLUBB).

The results in GrADS output files from this revision is an exact match to the
results in GrADS output files from the previous revision for all cases.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8349 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: e7fe6b09e563382cdf65378edd286fa3d6b9c7c9