I have added the coding infrastructure to output additional statistics w_1_zm, w_2_zm, varnce_w_1_zm, varnce_w_2_zm, and mixt_frac_zm.  These are necessary for CLUBB restarts if the ipdf_call_placement flag is set to ipdf_post_advance_fields and the l_call_pdf_closure_twice flag is turned on.
Additionally, I have added all the necessary CLUBB fields to input fields and
to CLUBB restarts so that restarts should be successful if the
ipdf_call_placement flag is set to ipdf_post_advance_fields.

I have not added the new stats to the stats.in files yet.  With that, the
results in GrADS output files from this revision are an exact match to the
results in GrADS output files from the previous revision for all cases.

Larson Group Internal Trac System Reference: cam:ticket:87.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8362 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 4289d2475f8d595b9e6dc2bb91286da307b49241