Now that the PDF component correlations of w and chi, as well as w and eta, will be allowed to be something other than 0 when a new PDF is used, stats needs to account for that.
Already in the code, corr_chi_eta_1 is the value calculated in PDF closure,
while corr_chi_eta_1_ca is the actual value used in the correlation array.
This depends on whether the calculated value or a prescribed value is used
in the array.  This code is now being mimicked for corr_w_chi_1, corr_w_chi_2,
corr_w_eta_1, and corr_w_eta_2.

The first step is change the current stats output corr_w_chi_1 to
corr_w_chi_1_ca, etc.

The results in GrADS *.dat output files from this revision are an exact match
to the results in GrADS *.dat output files from the previous revision for all
cases.  The GrADS *.ctl files differ because the stat names have changed.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8413 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 8a0b4efb9a160adb315c099462bcc07dc5b9359e