I am adding statistical output for corr_w_chi_1, corr_w_chi_2, corr_w_eta_1, and corr_w_eta_2.
The results in GrADS output files differ because of the additional statistics,
so this revision is BIT_CHANGING.  However, this is only for the *_zt output
files.  The results in GrADS *_zm and *_sfc output files from this revision are
and exact match to the results in GrADS *_zm and *_sfc files from the previous
revision for all cases.

CLUBB Trac ticket reference: clubb:ticket:806.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8414 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 309a5fed35104c5422cc4ad24efc8c0c241ad858