I have now switched the PDF closure code over to calling the code that had been moved to original_pdf.F90 in r8450.  The code that remained in PDF closure has now been deleted.
I have also moved the code to calculate the PDF parameters for passive scalars
to inside the driver subroutines for ADG1 and ADG2.

The results in GrADS output files from this revision are an exact match to the
results in GrADS output files from the previous revision for all cases.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8451 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 5b68933fc999bc8258625b0d9a3fd8c1b3128d46