I am adding a new function, calc_wpxpyp_pdf, to pdf_closure_module.F90.  This function calculates the <w'x'y'> term that I derived by integrating over the PDF.  This function is used to calculate <w'rt'thl'>.  Shortly, it will also be used to calculate <w'sclr'rt'> and <w'sclr'thl'>.  This new calculation of <w'x'y'> allows for the PDF component correlations involving w to have a value other than 0.  This generalizes the calculation for a PDF other than ADG1 or ADG2.
CLUBB's calculation of wprtpthlp in pdf_closure is currently just used for
statistical output.  For ADG1, the old form and the new form of wprtpthlp
produce the same result mathematically.  However, owing to numerical roundoff,
the results of the wprtpthlp in the new revision are BIT_CHANGING compared to
the old revision.  Otherwise, the results in GrADS output files for all other
outputs from this revision are an exact match to the results in GrADS output
files from the previous revision for all cases.

CLUBB Trac ticket reference: clubb:ticket:777 and clubb:ticket:806.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8546 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 85a115499facd5577647492c2a1cc8c1419b2426