I altered the names of the turbulent advection functions that are closed by the PDF (from integration over the PDF).  The current functions have been altered to include ADG1 in their names, because the closure is done by assuming ADG1 is the PDF used.
The results of this revision are an exact match to the results from the previous
revision for all cases.  However, yesterday's clubb:r8587 changed results
without including BIT_CHANGING in the commit message, so that is being added
here to reset the appropriate Bitten test.

CLUBB Trac ticket reference: clubb:ticket:804.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8589 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 59983e00cb8d7b0d64af979ab088c411ff760466