I am updating the code that advances predictive variances and covariances in advance_xp2_xpyp_module.F90 to use one set of turbulent advection functions for all turbulent advection options (explicit or implicit/semi-implicit based on PDF type, as well as centered or "upwind" discretization").
Along the way, I discovered that the current ADG1 implementation, which is
supposed to be using "upwind" discretization, is only using "upwind"
discretization for the implicit component of the turbulent advection term, while
using centered discretization for the explicit (RHS) component of the turbulent
advection term.  An "upwind" discretization for the explicit component was never
added to the code.  I assumed this was an oversight, and an "upwind"
discretization for the explicit component of the turbulent advection term was
added as part of the new code for this project (in a prior commit).  Both
implicit and explicit components now use either centered discretization or
"upwind" discretization consistently.

Owing to the RHS discretization correct, as well owing to a different order of
calculations, this commit is BIT_CHANGING.  However, plots of all cases show
that the previous revision and the current revision overlap for most cases.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8731 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: fdba5b894298fe351486792a92a521586dc1d9b7