I am updating the code that advances predictive scalar fluxes (<w'x'>) in advance_xm_wpxp_module.F90 to use one set of turbulent advection functions for all turbulent advection options (explicit or implicit/semi-implicit based on PDF type, as well as centered or "upwind" discretization).
For the default ADG1 PDF, the calculations are mathematically the same.
However, they occur in a different order in the new code.  This makes this
revision BIT_CHANGING.  However, plots of all cases show that the previous
revision and the current revision overlap for most cases.  Only the most
sensitive cases show differences on the plots.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8735 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: ef271ff2d4a0fb4b7ca7fdfbaba3a3d2f6cac016