I committed a subroutine, diagnose_upxp, that calculates upthlp, uprtp, vpthlp, vprtp using a simple diagnostic version of Eqn. 7 of Andre et al. (1978).  Ultimately, CLUBB will use this in order to compute the buoyancy terms in the upwp and vpwp budgets.
The results match the prior repo results bit for bit
if, in the prior code, I pull tau_C6_zm inside
the parentheses in the prior calculation of upthlp and uprtp:
upthvp = - 0.3_core_rknd * tau_C6_zm *    (    upwp * ddzt( thlm ) + wpthlp * ddzt( um ) &
         + 200.0_core_rknd * ( upwp * ddzt( rtm )  +  wprtp * ddzt( um ) ) ) &
This is needed to avoid a change in the order of operations, which changes
the results in IEEE arithmetic.  I also need to set C6=1 and C7=0 in
subroutine diagnose_upxp.

However, what I am committing includes C6 and C7 in subroutine diagnose_upxp.
It doesn't change the results much, but nevertheless the results are BIT_CHANGING.
In addition, four new stats are added to all_stats.in and standard_stats.in:
upthlp, uprtp, vpthlp, and vprtp.

For #841.