I am correcting the code so that the Nightly Uninitialized Variables Test (Oracle Fortran) works correctly again.
I've noticed that FORTRAN's where statement doesn't always prevent a certain
part of a where ... elsewhere block from being entered in the same way that it
prevents a certain part of an if ... elseif block from being entered.  Code that
is not supposed to be entered under certain conditions gets entered and then
the mask is applied later, throwing away the results.  However, if the
intrusion causes a floating point exception for whatever reason, the run stops.

I remember overhearing talk between Gunther and Vince about the where statement
"guessing" wrong sometimes.  Is there a way to stop it from "guessing" wrong
or at least reduce the number of wrong "guesses"?