I am making the first step for updating the code that calculates sigma_sqd_w. I have included upwp^2/(up2*wp2) and vpwp^2/(vp2*wp2) in the calculation of the maximum squared correlation for the limit of sigma_sqd_w when the l_predict_upwp_vpwp flag is enabled.
When the l_predict_upwp_vpwp flag is turned off, the results in GrADS output
files from this revision are an exact match to the results in GrADS output files
from the previous revision for all cases.

When the l_predict_upwp_vpwp flag in enabled (which it is by default), this
revision is BIT_CHANGING.  However, only a couple of cases show visible changes
on the plots.

Git Issue Reference: #852.