I switched on direct diagnosis of tau rather than Lscale. That is, I set l_diag_Lscale_from_tau = .true. Thus, this commit is BIT_CHANGING.
Because diagnosing tau changes answers, I retuned CLUBB,
taking the opportunity to add some features seem better justified.
For instance, we now damp wp2 using tke (l_damp_wp2_using_em = .true.)
because doing so is more standard in turbulence modeling and because
it may help reduce damping on Sc clouds.
C7 is allowed to become larger in statically stable layers
(l_use_C7_Richardson = .true.)
because that is what the LES of Heinze show, and because doing so
might help limit gravity waves above cloud.

Many parameters have been retuned.  I set C2rtthl = C2rt = C2thl
because it seems more justifiable and because doing so
produces the correct sign of wprtp_bp in Cu layers.
Also, the "b" parameters have been set equal to the non-"b"
parameters in order to reduce the number of tunable parameters.
The "b" parameters were based on skewness, which is hard to control and tune.
Instead, with this revised version, the tuning parameters are based mostly
on thlm, which is more robust.  I set C15=0 because this term
is complex and hard to justify.  c_K8 was increased because
it smooths the solutions without dominating the wp3 budget.
beta was reduced in order to make wp2xp larger and wpxp2 smaller.

I included a bug fix to Cx_fnc_Richardson, which prevents
C7 and C11 from exceeding the maximum set by Cx_max.

The brunt vaisala frequency is reduced for slightly stable layers
in order to allow Sc layers to remain well mixed.

Further tuning is needed to reduce oscillations and temporal noise.
Also, the momentum fields --- upwp, vpwp, up2, and vp2 ---
need to be retuned.  Otherwise, the results are comparable to the prior
default.

We're expecting a speed-up of 10% to 15% because computing CLUBB's
mixing length directly is so costly.

For #842.