I am committing a test where I changed subroutine ADG1_w_closure from using nested where statements to using if statements within the context of an elemental subroutine.  My notes are below:
===============================================================================

Here's an intriguing idea to try in place of nested "where" statements.

The article at:

https://data-science-austria.at/2018/12/24/the-need-for-speed-part-2-c-vs-fortran-vs-c/

says that using FORTRAN ELEMENTAL subroutines and functions is just as fast as
using FORTRAN serial loops with -O3 optimization (which is what we've been doing
for optimization).

The use of elemental subprograms contains many advantages:

1) The subprogram becomes much more versatile, because it can be used for
scalars as well as arrays of varying sizes and dimensions.

2) In our case, the changesets become easier to follow, as they are basically
a change from where statements to if ... then statements.  There are no
additional do loops and loop indices needed.  These additional statements and
indices can be confusing when comparing code differences and they clutter-up
(a.k.a. uglify) subprograms.

3) The article states that this method should be just as fast for vectors of
size 1000 and less.  For CLUBB, this would be the number of height levels,
which don't normally exceed 128.  I tested the old version (where statements)
and the new version for multiple runs of BOMEX and ARM, and the timings were
equivalent.

4) This avoids the problem of code entering the portion of a "where" statement
that it shouldn't be and crashing when floating point trapping is enabled.

The use of elemental subprograms has the following limitations:

1) All input and output variables must be of the same size and dimension.

In other words, whatever code is calling the elemental subroutine must feed
it, as well as take as output, either all scalar values, or all vectors of
size nz, or all 2-D arrays of size nx by ny, etc.

2) Elemental subprograms are also pure subprograms, so they have all the
restrictions of pure subprograms (no print statements, etc.).

================================================================================

The results in GrADS output files from this revision are an exact match to the
results in GrADS output files from the previous revision for all cases.

Obligatory Larson Group GitHub Issue Reference: #859.