I have resolved all the issues surrounding recent CLUBB commits that remove "use" statements from the code.  The commit 068a7b2 (April 8, 2020) was the commit before these refactorings started.  When Lscale, Skw_zm, and thvm are passed out of CLUBB core, the code returns to a bit-for-bit match to commit 068a7b2.  However, rather than pass these variables out of advance_clubb_core, it is more convenient to recalculate them locally in clubb_driver.  This changes results for (a) SILHS cases, owing to the change to Lscale, and (b) all cases that advance microphysics, owing to the change to Skw_zm.  The net result is that this commit is BIT_CHANGING.  For cases that advance microphysics and/or use SILHS, the net result is BIT_CHANGING when compared to 068a7b2.  For all other cases (BOMEX, etc.), the results are an exact match.
It is important to note that host models that use CLUBB (without SILHS) will not
be affected by these changes.  Host models that use SILHS will only be affected
by the change to Lscale.  However, this code has already been applied to CAM and
E3SM, so the only changes will to SILHS code in SAM and WRF.

Larson Group Issue Reference: #876.