I am making the change of using tau_wp2_zm to damp up2 and vp2, rather than tau_xp2_zm, when l_diag_Lscale_from_tau is enabled. The previous code was causing up2 and vp2 to become large and unstable.
Additionally, I placed a limiter on tau based on Lscale_max in
the CLUBB code for the situation where l_diag_Lscale_from_tau
is enabled.  This will help constrict tau, and thereby the
magnitudes of turbulent fields, in a host model as grid spacing
starts to become arbitrarily small and resolved features take
over.

This revision is BIT_CHANGING.