I am changing the new hybrid PDF's calculation of F_w to use tunable parameter gamma (technically, the gamma_Skw_fnc).
In the ADG1 PDF, when gamma_Skw_fnc goes to 0, the standard
deviations of w in each PDF component both go to 0, and the
spread between the PDF component means of w is at its maximum.
When gamma_Skw_fnc goes to 1, the standard deviations of w in
each PDF component become large, and the spread between the PDF
component means of w becomes small.

The new hybrid PDF is based on the same concept, except that I
had been using a hard-wired 0.75 in the place of gamma or the
gamma_Skw_fnc.  Re-using the tunable parameters for gamma here
makes sense, since it's the same concept and this makes tuning
easier.