I have pared two subprograms down to one again, this time for the subprograms found in module diffusion.
As part of this, the calls to diffusion (for turbulent advection)
in advance_microphys_module.F90 needed to be upgraded to use the
fully optimized version of the code. As a result, this revision
is BIT_CHANGING, but only for cases that use microphysics. A print
out of the LHS array from the call to diffusion shows that numbers
are in agreement until you reach the point of numerical roundoff.

CLUBB Git Issue Reference: #878.