I added altitude_threshold to the list of CLUBB's tunable parameters.
The results in GrADS output files from this revision are an exact match to the
results in GrADS output files from the previous revision for all cases.

However, NetCDF output files are not an exact match, owing to the presence
of the new tunable parameter listed in the output files.  That technically
makes this revision BIT_CHANGING.