These updates incorporate some of Shixuan Zhang's changes, including:
(1) upwind scheme for the eddy diffusion code, which Shixuan says does not make a huge difference; and
(2) l_smooth_heaviside_tau_wpxp flag in the mixing_length.F90 code, which is intended to give
smoother results.

Both of the flags for these changes are turned off, but these updates are BIT_CHANGING because
of some minor changes to the way that factors of rho_ds_* and invrs_rho_ds_* are handled going
into the diffusion.F90 subroutines.  Most cases look identical in pyplotgen, but sensitive cases
like arm_97 and rico look slightly different.

These changes were originally in pull request #962, but because of conflicts with the master
branch code base that developed, they could not be rebased and merged, so I incorporated them
into this new commit.