I altered the w_up_in_cloud and w_down_in_cloud code so that a thresholding is used similar to what is seen in the cloud fraction and cloud water code.  If the PDF component mean of w more than the maximum allow number of standard deviations away from 0, the PDF component is either all-updrafty or all-downdrafty, and the code avoids expensive computations where large magnitude values can potentially be fed into ERF or EXP.
Of course, the results are not bit-for-bit with the previous version,
meaning that there are some situations where these thresholds come
into effect.  However, plots of all cases show no visible differences
in w_up_in_cloud and w_down_in_cloud.

Since w_up_in_cloud and w_down_in_cloud are not included in normal
stats output, this commit does not change the bit-for-bitness of
CLUBB code.